pyproject.toml
[tool.poetry]
name = "img-concat"
version = "0.1.0"
description = ""
authors = ["HaroldRoot <ngoksausi618@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pillow = "^11.0.0"
colorama = "^0.4.6"
rich = "^13.9.4"
imagehash = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
